Add functions to get names for type or member.
Now able to specify the retrieval of element type and member names and element type file name or member file name.

A type file name is the proto file in which an element's type was defined. Note that this will be an empty string for built in types.

A member file is the proto file in which an element member was used.